requirements.txt
requests==2.25.0
fastapi==0.61.2
uvicorn==0.12.3
gunicorn==20.0.4
behave==1.2.6
jinja2==2.11.3
aiofiles==0.6.0
coverage==5.3
pytest==6.1.2
pytest-cov==2.10.1
pytest-flakes==4.0.2
pytest-pep8==1.0.6
pytest-pythonpath==0.7.3
pyyaml==5.3.1
pycodestyle==2.6.0
pymongo
pydantic
